Revert "Merge pull request #12 from MatheusHNSantos/Matheus"
This reverts commit 598009846581a9fe33c9f87101b047bc6b38323f, reversing
changes made to d4fedfa239f0c7628440cfe58a393411a765d2ac.
