+++ nbproject/jfx-impl.xml
@@ -327,26 +327,48 @@
         <property name="ant-javafx.jar.location" value="${platform.home}${file.separator}tools${file.separator}ant-javafx.jar"/>
     </target>
     <target name="-check-ant-javafx-in-probjdk-lib" unless="ant-javafx.jar.location">
+        <condition property="do.set.ant-javafx.in.probjdk.lib.has_jre">
+            <available file="${java.home}${file.separator}..${file.separator}lib${file.separator}ant-javafx.jar"/>
+        </condition>
+        <condition property="do.set.ant-javafx.in.probjdk.lib.no_jre">
+            <available file="${java.home}${file.separator}lib${file.separator}ant-javafx.jar"/>
+        </condition>
         <condition property="do.set.ant-javafx.in.probjdk.lib">
             <and>
                 <not><isset property="ant-javafx.jar.location"/></not>
-                <available file="${java.home}${file.separator}..${file.separator}lib${file.separator}ant-javafx.jar"/>
+                <or>
+                    <isset property="do.set.ant-javafx.in.probjdk.lib.has_jre"/>
+                    <isset property="do.set.ant-javafx.in.probjdk.lib.no_jre"/>
+                </or>
             </and>
         </condition>
     </target>
     <target name="-set-ant-javafx-in-probjdk-lib" depends="-set-ant-javafx-in-platform-home-tools,-check-ant-javafx-in-probjdk-lib" if="do.set.ant-javafx.in.probjdk.lib">
-        <property name="ant-javafx.jar.location" value="${java.home}${file.separator}..${file.separator}lib${file.separator}ant-javafx.jar"/>
+        <condition property="ant-javafx.jar.location" value="${java.home}${file.separator}..${file.separator}lib${file.separator}ant-javafx.jar" else="${java.home}${file.separator}lib${file.separator}ant-javafx.jar">
+            <isset property="do.set.ant-javafx.in.probjdk.lib.has_jre"/>
+        </condition>
     </target>
     <target name="-check-ant-javafx-in-probjdk-tools" unless="ant-javafx.jar.location">
+        <condition property="do.set.ant-javafx.in.probjdk.tools.has_jre">
+            <available file="${java.home}${file.separator}..${file.separator}tools${file.separator}ant-javafx.jar"/>
+        </condition>
+        <condition property="do.set.ant-javafx.in.probjdk.tools.no_jre">
+            <available file="${java.home}${file.separator}tools${file.separator}ant-javafx.jar"/>
+        </condition>
         <condition property="do.set.ant-javafx.in.probjdk.tools">
             <and>
                 <not><isset property="ant-javafx.jar.location"/></not>
-                <available file="${java.home}${file.separator}..${file.separator}tools${file.separator}ant-javafx.jar"/>
+                <or>
+                    <isset property="do.set.ant-javafx.in.probjdk.tools.has_jre"/>
+                    <isset property="do.set.ant-javafx.in.probjdk.tools.no_jre"/>
+                </or>
             </and>
         </condition>
     </target>
     <target name="-set-ant-javafx-in-probjdk-tools" depends="-set-ant-javafx-in-probjdk-lib,-check-ant-javafx-in-probjdk-tools" if="do.set.ant-javafx.in.probjdk.tools">
-        <property name="ant-javafx.jar.location" value="${java.home}${file.separator}..${file.separator}tools${file.separator}ant-javafx.jar"/>
+        <condition property="ant-javafx.jar.location" value="${java.home}${file.separator}..${file.separator}tools${file.separator}ant-javafx.jar" else="${java.home}${file.separator}tools${file.separator}ant-javafx.jar">
+            <isset property="do.set.ant-javafx.in.probjdk.tools.has_jre"/>
+        </condition>
     </target>
     <target name="-check-ant-javafx-in-macjdk-lib" unless="ant-javafx.jar.location">
         <condition property="do.set.ant-javafx.in.macjdk.lib">
@@ -1825,6 +1847,25 @@
                     
                     // manifest
                     var man = jar.createManifest();
+                    var userManifestPath = project.getProperty("manifest.file");
+                    if (userManifestPath) {
+                        var userManifestFile = project.resolveFile(userManifestPath);
+                        if (userManifestFile.isFile()) {
+                            var manifestEncoding = project.getProperty("manifest.encoding");
+                            var userManifestReader = manifestEncoding ?
+                                    new java.io.InputStreamReader(
+                                        new java.io.FileInputStream(userManifestFile),
+                                        manifestEncoding) :
+                                    new java.io.InputStreamReader(
+                                        new java.io.FileInputStream(userManifestFile));
+                            try {
+                                var userManifest = new org.apache.tools.ant.taskdefs.Manifest(userManifestReader);
+                                man.merge(userManifest);
+                            } finally {
+                                userManifestReader.close();
+                            }
+                        }
+                    }
                     var a1val = new String(project.getProperty("application.vendor"));
                     var a1 = new org.apache.tools.ant.taskdefs.Manifest.Attribute();
                     a1.setName("Implementation-Vendor");
