+++ nbproject/private/configs/Executar_no_Browser.properties
@@ -1,4 +1,4 @@
 # Do not modify this property in this configuration. It can be re-generated.
 javafx.run.as=embedded
 javafx.run.inbrowser=<Browser Default do Sistema>
-javafx.run.inbrowser.path=C:\\Program Files (x86)\\baidu\\Baidu Browser\\Spark.exe
+javafx.run.inbrowser.path=C:\\Program Files\\Internet Explorer\\IEXPLORE.EXE
