+++ nbproject/private/private.properties
@@ -1,4 +1,4 @@
 auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
 # N\u00e3o h\u00e1 necessidade de se alterar essa propriedade, a menos que se esteja personalizando a infraestrutura da tarefa Ant do JavaFX
 endorsed.javafx.ant.classpath=.
-user.properties.file=C:\\Users\\mathe_000\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
+user.properties.file=/home/felipemantoan/.netbeans/8.2/build.properties
