+++ src/model/User.java
@@ -16,7 +16,7 @@ public class User extends Employee implements ModelInterface{
     private String login;
     private String password;
    
-    public User(String login, String password) {
+    User(String login, String password) {
         this.login = login;
         this.password = password;
     }
